Added a new folder around getting InGameAnnouncement
