+++ Eevee.Sleep.Bot/Controllers/Mongo/InGameAnnouncement/InGameAnnouncementIndexController.cs
@@ -1,7 +1,7 @@
-using Eevee.Sleep.Bot.Models;
+using Eevee.Sleep.Bot.Models.InGameAnnouncement;
 using MongoDB.Driver;
 
-namespace Eevee.Sleep.Bot.Controllers.Mongo;
+namespace Eevee.Sleep.Bot.Controllers.Mongo.InGameAnnouncement;
 
 public static class InGameAnnouncememntIndexController {
     public static Task BulkUpsert(InGameAnnouncementIndexModel[] model) {
