+++ Eevee.Sleep.Bot/Controllers/Mongo/MongoConst.cs
@@ -1,5 +1,6 @@
 ﻿using Eevee.Sleep.Bot.Extensions;
 using Eevee.Sleep.Bot.Models;
+using Eevee.Sleep.Bot.Models.InGameAnnouncement;
 using Eevee.Sleep.Bot.Utils;
 using MongoDB.Driver;
