+++ Eevee.Sleep.Bot/Models/InGameAnnouncement/InGameAnnouncementIndexModel.cs
@@ -1,6 +1,6 @@
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace Eevee.Sleep.Bot.Models;
+namespace Eevee.Sleep.Bot.Models.InGameAnnouncement;
 
 [BsonIgnoreExtraElements]
 public record InGameAnnouncementIndexModel: InGameAnnouncementMetaModel;
